==> R/00_doit.R <==
# Load data
source("R/01_load.R")
# Clean it up
source("R/02_clean.R")
# Prepare data for analysis
source("R/03_augment.R")

# Do analysis
source("R/04_analysis_i.R")
source("R/05_analysis_ii.R")
source("R/06_analysis_iii.R")
source("R/07_analysis_iv.R")
source("R/08_analysis_v.R")

rm(list=ls())
